Backend/chatapp/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
django-heroku==0.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
